+++ gsad/src/gsad_http_handler.c
@@ -30,6 +30,7 @@
 #include "gsad_i18n.h"        /* for accept_language_to_env_fmt */
 #include "gsad_settings.h"    /* for get_guest_usernmae */
 #include "gsad_user.h"        /* for user_t */
+#include "utils.h"            /* for str_equal */
 #include "validator.h"        /* for gvm_validate */
 
 #include <assert.h>              /* for assert */
@@ -412,7 +413,6 @@ handle_setup_user (http_connection_t *connection, const char *method,
 
   if (ret == USER_GMP_DOWN)
     {
-
       return handler_send_reauthentication (
         connection, MHD_HTTP_SERVICE_UNAVAILABLE, GMP_SERVICE_DOWN);
     }
@@ -688,7 +688,7 @@ handle_static_file (http_connection_t *connection, const char *method,
   g_debug ("Requesting url %s for static ng path %s", url, path);
 
   response_data = cmd_response_data_new ();
-  cmd_response_data_set_allow_caching (response_data, 1);
+  cmd_response_data_set_allow_caching (response_data, TRUE);
 
   response = file_content_response (connection, url, path, response_data);
 
@@ -717,7 +717,9 @@ handle_static_config (http_connection_t *connection, const char *method,
   g_debug ("Requesting url %s for static config path %s", url, path);
 
   response_data = cmd_response_data_new ();
-  cmd_response_data_set_allow_caching (response_data, 1);
+
+  // don't cache config file
+  cmd_response_data_set_allow_caching (response_data, FALSE);
 
   if (g_file_test (path, G_FILE_TEST_EXISTS))
     {
@@ -756,11 +758,12 @@ init_http_handlers ()
 
   url_handlers = url_handler_new ("^/(img|js|css|locales)/.+$",
                                   http_handler_new (handle_static_file));
-  url_handler_add_func (url_handlers, "^/robots.txt$", handle_static_file);
+  url_handler_add_func (url_handlers, "^/robots\\.txt$", handle_static_file);
 
-  url_handler_add_func (url_handlers, "^/config.*js$", handle_static_config);
+  url_handler_add_func (url_handlers, "^/config\\.*js$", handle_static_config);
   url_handler_add_func (url_handlers, "^/static/(img|js|css)/.+$",
                         handle_static_file);
+  url_handler_add_func (url_handlers, "^/manual/.+$", handle_static_file);
 
   http_handler_t *gmp_handler = http_handler_new (handle_setup_user);
   http_handler_add (gmp_handler, http_handler_new (handle_setup_credentials));
@@ -832,12 +835,14 @@ handle_request (void *cls, http_connection_t *connection, const char *url,
   gsad_connection_info_t *con_info;
   http_handler_t *handlers;
   char *full_url;
+  gboolean new_connection;
 
   handlers = (http_handler_t *) cls;
   con_info = *con_cls;
+  new_connection = (*con_cls == NULL);
 
   /* Never respond on first call of a GET. */
-  if ((!strcmp (method, "GET")) && *con_cls == NULL)
+  if ((str_equal (method, "GET")) && new_connection)
     {
       /* First call for this request, a GET. */
 
@@ -855,9 +860,9 @@ handle_request (void *cls, http_connection_t *connection, const char *url,
       return MHD_YES;
     }
 
-  if (!strcmp (method, "POST"))
+  if (str_equal (method, "POST"))
     {
-      if (NULL == *con_cls)
+      if (new_connection)
         {
           /* First call for this request, a POST. */
 
@@ -866,7 +871,8 @@ handle_request (void *cls, http_connection_t *connection, const char *url,
 
           con_info->postprocessor = MHD_create_post_processor (
             connection, POST_BUFFER_SIZE, serve_post, (void *) con_info);
-          if (NULL == con_info->postprocessor)
+
+          if (con_info->postprocessor == NULL)
             {
               g_free (con_info);
               /* Both bad request or running out of memory will lead here, but
@@ -877,6 +883,7 @@ handle_request (void *cls, http_connection_t *connection, const char *url,
                              GSAD_CONTENT_TYPE_TEXT_HTML, NULL, 0);
               return MHD_YES;
             }
+
           con_info->params = params_new ();
           con_info->connectiontype = 1;
 
@@ -886,7 +893,7 @@ handle_request (void *cls, http_connection_t *connection, const char *url,
 
       /* Second or later call for this request, a POST. */
 
-      if (0 != *upload_data_size)
+      if (*upload_data_size != 0)
         {
           MHD_post_process (con_info->postprocessor, upload_data,
                             *upload_data_size);