+++ CHANGELOG.md
@@ -22,6 +22,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Add tooltips to deactivated text fields in AlertDialog [#1269](https://github.com/greenbone/gsa/pull/1269)
 
 ### Changed
+- Avoid storing config.js in browser cache [#1372](https://github.com/greenbone/gsa/pull/1372)
 - Display data loading errors at list pages [#1349](https://github.com/greenbone/gsa/pull/1349)
 - Improve login page [#1347](https://github.com/greenbone/gsa/pull/1347)
 - Remove options for not implemented languages [#1344](https://github.com/greenbone/gsa/pull/1344)
@@ -38,6 +39,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Cleanup get_report function in gsad [#1263](https://github.com/greenbone/gsa/pull/1263)
 
 ### Fixed
+- Use German manual for *DE* locale [#1372](https://github.com/greenbone/gsa/pull/1372)
 - Load all container tasks for report import dialog from redux store [#1370](https://github.com/greenbone/gsa/pull/1370)
 - Don't render *Invalid Date* [#1368](https://github.com/greenbone/gsa/pull/1368)
 - Don't show error message after re-login [#1366](https://github.com/greenbone/gsa/pull/1366)